Successfully added hotkey support
The 'keyboard' library I found in PyPI is superb! It makes it easy for a
developer to implement multi key combination even if your main
application is out of focus. This new hotkey feature is now implemented
in Tenny. CTRL+SHIFT+S starts and stops the timer while CTRL+SHIFT+R
resets the timer.