Successfully implemented amending the hotkey!
Wow! It's been a while since my last commit. I was struggling on
amending the hotkeys while the application is running and designing how
will the user amend the hotkeys. It was a tough fortnight so I needed to
have a break to refresh my head to think for other solutions. I miss
commiting especially seeing those contribution chart!!